+++ blocks/shared-styles/_children/promos/large/mock-data-video.js
